model/requirements.txt
tensorflow==2.15.0.post1
keras==2.15.0
numpy==1.26.1
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
scikit-learn==1.3.2
pika==1.3.2
python-dotenv==1.0.0
